Fix "no user with id [...]" problem in partner channels
When using partner channels (i.e. [cross-org
channels][partner-channels]), the [`users.list`][users-list] request
might not return enough information to fill in the user names of the
conversation.

Let's use the (slow!) route of [`users.info`][users-info] as a fall-back
whenever a user ID is encountered that finds no match in the result of
[`users.list`][users-list].

[partner-channels]: https://slack.com/resources/using-slack/introducing-your-partners-to-slack-connect
[users-list]: https://api.slack.com/methods/users.list
[users-info]: https://api.slack.com/methods/users.info