feat(command): add api command
This adds a more generic way to interact with the slack API, it is a
one-stop method that _could_ support any request to the slack API. This
means the existing client.get/client.post could be rewritten to use this
method.

Adding an API handler goes very much in the `gh api` way as well.
Sometimes the helpers aren't enough and proxying to the API is the only
real way to make something.

This will need further consideration and documentation writing to
explain the different cases and gotchas of using the API directly.